+++ src/onchain/pluts/PTypes/PData/conversion/__tests__/PTypes.PData.conversion.getFromDataForType.test.ts
@@ -1,0 +1,14 @@
+import { getFromDataForType } from "../getFromDataTermForType"
+import PValue from "../../../../API/V1/Value/PValue"
+
+describe("getFromDataForType", () => {
+
+    test("PValue", () => {
+        
+        expect(
+            () => getFromDataForType( PValue.type )
+        ).not.toThrow();
+
+    })
+
+})