+++ src/onchain/pluts/PTypes/PPair/index.ts
@@ -1,16 +1,17 @@
-import JsRuntime from "../../../utils/JsRuntime";
-import PType, { PDataRepresentable } from "../PType";
-import { phoist, plam } from "../Syntax/syntax";
-import Term from "../Term";
-import Type, { ConstantableTermType, data, DataType, pair, TermType, ToPType } from "../Term/Type/base";
-import { typeExtends } from "../Term/Type/extension";
-import PData from "./PData/PData";
-import PLam from "./PFn/PLam";
-import { punsafeConvertType } from "../Syntax";
-import { isConstantableTermType } from "../Term/Type/kinds";
-import UPLCConst from "../../UPLC/UPLCTerms/UPLCConst";
-import { termTyToConstTy } from "../Term/Type/constTypeConversion";
-import TermPair, { addPPairMethods } from "../stdlib/TermPair";
+import JsRuntime from "../../../../utils/JsRuntime";
+import PType, { PDataRepresentable } from "../../PType";
+import { phoist, plam } from "../../Syntax/syntax";
+import Term from "../../Term";
+import Type, { ConstantableTermType, data, DataType, pair, TermType, ToPType } from "../../Term/Type/base";
+import { typeExtends } from "../../Term/Type/extension";
+import PData from "../PData/PData";
+import PLam from "../PFn/PLam";
+import { punsafeConvertType } from "../../Syntax";
+import { isConstantableTermType } from "../../Term/Type/kinds";
+import UPLCConst from "../../../UPLC/UPLCTerms/UPLCConst";
+import { termTyToConstTy } from "../../Term/Type/constTypeConversion";
+import TermPair, { addPPairMethods } from "../../stdlib/TermPair";
+import ObjectUtils from "../../../../utils/ObjectUtils";
 
 export default class PPair<A extends PType, B extends PType > extends PDataRepresentable
 {