+++ src/onchain/pluts/Syntax/punsafeConvertType.ts
@@ -1,28 +1,34 @@
+import BasePlutsError from "../../../errors/BasePlutsError";
 import ObjectUtils from "../../../utils/ObjectUtils";
 import PType from "../PType";
 import Term from "../Term";
 import { TermType, ToPType } from "../Term/Type/base";
+import { isWellFormedType } from "../Term/Type/kinds";
 
 
 export default function punsafeConvertType<FromPInstance extends PType, SomeExtension extends {}, ToTermType extends TermType>
 ( someTerm: Term<FromPInstance> & SomeExtension, toType: ToTermType ): Term<ToPType<ToTermType>> & SomeExtension
 {
-const converted = new Term(
-    toType,
-    someTerm.toUPLC
-) as any;
+    if( !isWellFormedType( toType ) )
+    throw new BasePlutsError("");
 
-Object.keys( someTerm ).forEach( k => {
+    const converted = new Term(
+        toType,
+        someTerm.toUPLC,
+        Boolean((someTerm as any).isConstant) // isConstant
+    ) as any;
 
-    if( k === "_type" || k === "_toUPLC" ) return;
-    
-    ObjectUtils.defineReadOnlyProperty(
-        converted,
-        k,
-        (someTerm as any)[ k ]
-    )
+    Object.keys( someTerm ).forEach( k => {
 
-});
+        if( k === "_type" || k === "_toUPLC" ) return;
+        
+        ObjectUtils.defineReadOnlyProperty(
+            converted,
+            k,
+            (someTerm as any)[ k ]
+        )
 
-return converted;
+    });
+
+    return converted;
 }